+++ packages/api/app/distribution-routes.js
@@ -11,6 +11,8 @@ const {
 const displayS3CredentialInstructions = require('../endpoints/s3credentials-readme');
 const { isAccessTokenExpired } = require('../lib/token');
 
+const version = require('../endpoints/version');
+
 /**
  * Helper function to pull bucket out of a path string.
  * Will ignore leading slash.
@@ -106,8 +108,6 @@ const pubkey = (req, res) => res.status(501).end();
 
 const s3Credentials = (req, res) => res.status(501).end();
 
-const version = (req, res) => res.status(501).end();
-
 router.get('/', root);
 router.get('/locate', locate);
 router.get('/login', login);
@@ -117,7 +117,8 @@ router.get('/pubkey', pubkey);
 router.get('/redirect', handleRedirectRequest);
 router.get('/s3Credentials', s3Credentials);
 router.get('/s3CredentialsREADME', displayS3CredentialInstructions);
-router.get('/version', version);
+// Use router.use to leverage custom version middleware
+router.use('/version', version);
 
 // HEAD /*
 // GET /* <- Actual presigned URL